Refactor car demolition logic
